fix: Initial commit
Couple of simple scripts to get some info from the text exported from
RegEdit.exe.
They are just tidied copies made from from REPL exploration of the files
I was doing when I was trying to figure out what to expect from reading
the Win Reg. Not really useful, but worth keeping and would not be
included in a build.